refactor(nms): Move sequelize-models out of fbc_js_core
