+++ nms/.eslintrc.js
@@ -259,7 +259,6 @@ module.exports = {
         'jest.config.js',
         'jest.*.config.js',
         'fbc_js_core/platform_server/**/*.js',
-        'fbc_js_core/sequelize_models/**/*.js',
         'fbc_js_core/util/yarn.js',
         'fbc_js_core/webpack_config/**/*.js',
         'config/*.js',
