+++ nms/scripts/createOrganization.js
@@ -15,12 +15,12 @@
  */
 'use strict';
 
-import type {OrganizationType} from '../fbc_js_core/sequelize_models/models/organization';
+import type {OrganizationType} from '../shared/sequelize_models/models/organization';
 import type {Tab} from '../shared/types/tabs';
 
 import Sequelize from 'sequelize';
 
-import {Organization} from '../fbc_js_core/sequelize_models';
+import {Organization} from '../shared/sequelize_models';
 import {coerceToTab} from '../shared/types/tabs';
 import {difference} from 'lodash';
 import {getProjectTabs} from '../shared/projects/projects';
