+++ nms/server/apicontroller/auditLoggingDecorator.js
@@ -20,7 +20,7 @@ import type {FBCNMSRequest} from '../auth/access';
 const url = require('url');
 import pathToRegexp from 'path-to-regexp';
 
-import {AuditLogEntry} from '../../fbc_js_core/sequelize_models';
+import {AuditLogEntry} from '../../shared/sequelize_models';
 const logger = require('../../shared/logging').getLogger(module);
 
 const defaultResolver = (req: FBCNMSRequest, type: string) => {
