+++ nms/server/auth/passport.js
@@ -15,10 +15,10 @@
  */
 
 import passport from 'passport';
-import {User} from '../../fbc_js_core/sequelize_models';
+import {User} from '../../shared/sequelize_models';
 
 import type {FBCNMSMiddleWareRequest} from '../../server/middleware';
-import type {UserType} from '../../fbc_js_core/sequelize_models/models/user';
+import type {UserType} from '../../shared/sequelize_models/models/user';
 
 type OutputRequest<T> = {
   logIn: (T, (err?: ?Error) => void) => void,
