+++ nms/server/auth/strategies/OrganizationLocalStrategy.js
@@ -15,7 +15,7 @@
  */
 
 import type {FBCNMSMiddleWareRequest} from '../../../server/middleware';
-import type {UserType} from '../../../fbc_js_core/sequelize_models/models/user.js';
+import type {UserType} from '../../../shared/sequelize_models/models/user.js';
 
 import bcrypt from 'bcryptjs';
 import {Strategy as LocalStrategy} from 'passport-local';
