+++ nms/server/auth/strategies/OrganizationSamlStrategy.js
@@ -16,7 +16,7 @@
 
 import {AccessRoles} from '../../../shared/roles';
 import {MultiSamlStrategy} from 'passport-saml'; // compatibility with breaking change in 3.1.0
-import {User} from '../../../fbc_js_core/sequelize_models';
+import {User} from '../../../shared/sequelize_models';
 
 import {getUserFromRequest} from '../util';
 import {injectOrganizationParams} from '../organization';
