+++ nms/server/features.js
@@ -18,7 +18,7 @@ import type {ExpressRequest, ExpressResponse, NextFunction} from 'express';
 import type {FBCNMSRequest} from './auth/access';
 import type {FeatureID} from '../shared/types/features';
 
-import {FeatureFlag} from '../fbc_js_core/sequelize_models';
+import {FeatureFlag} from '../shared/sequelize_models';
 
 export type RequestInfo = {
   isDev: boolean,
