+++ nms/shared/sequelize_models/README.md
