+++ nms/shared/sequelize_models/models/user.js
