+++ nms/shared/sequelize_models/sequelizeConfig.js
