+++ nms/shared/sequelize_models/sequelizerc.js
